+++ ci/authn-k8s/test.sh
@@ -75,7 +75,7 @@ function createNginxCert() {
 
 function buildDockerImages() {
   conjur_version=$(echo "$(< ../../VERSION)-$(git rev-parse --short=8 HEAD)")
-  DOCKER_REGISTRY_PATH="registry.tld/test"
+  DOCKER_REGISTRY_PATH="registry.tld"
 
   docker pull $DOCKER_REGISTRY_PATH/conjur:$conjur_version
   docker pull $DOCKER_REGISTRY_PATH/conjur-test:$conjur_version