+++ push-image.sh
@@ -1,76 +1,106 @@
 #!/usr/bin/env bash
-set -e
+set -euo pipefail
 
-# Push the 'conjur' image to various Docker registries
-# Push tagged images on master branch
-# Release images can be created by passing the desired tag to this script
-# Ex: ./push-image 4.9.5.1
+# Pushes the 'conjur' image to various Docker registries
+# There are essentially three flows this script supports
+# - Pushing images to the internal registry during the early pipeline stage
+# - Pushing an "edge"-tagged image to DockerHub
+# - Publishing images to public registries on new project releases during a
+#   tag-triggered build
+#
+# The default flow when no parameters are supplied is to publish images to
+# public registries. This does require that the TAG_NAME variable is set, which
+# happens automatically in tag-triggered builds.
+#
+# To publish to a (private) internal registry, run the script with the
+# "--registry-prefix=*" flag to supply the registry prefix.
+#
+# To publish an "edge"-tagged release to DockerHub, run the script with the
+# "--edge" flag.
 
 # shellcheck disable=SC1091
-. version_utils.sh
+. build_utils.sh
 
-if [[ -z "${TAG_NAME:-}" ]]; then
-  echo "Please supply environment variable TAG_NAME."
-  echo "If you see this error in Jenkins it means the publish script was run"
-  echo "for a build that wasn't triggered by a tag -" \
-    "please check publish stage conditions."
-  exit 1
-fi
+function print_help() {
+  echo "Usage: $0 [OPTION...]"
+  echo " --registry-prefix=STRING; optional registry prefix to prepend to '/cyberark/conjur'"
+  echo " --edge; optional flag to publish edge-tagged image to DockerHub"
+}
 
-TAG="${1:-$(version_tag)}"
-VERSION="$(< VERSION)"
-SOURCE_IMAGE="conjur:$TAG"
-RH_SOURCE_IMAGE="conjur-ubi:$TAG"
-
-CONJUR_REGISTRY="registry.tld"
+REGISTRY_PREFIX=""
+PUBLISH_EDGE=false
+TAG="$(version_tag)"
+LOCAL_IMAGE="conjur:${TAG}"
+RH_LOCAL_IMAGE="conjur-ubi:${TAG}"
 IMAGE_NAME="cyberark/conjur"
 REDHAT_IMAGE="scan.connect.redhat.com/ospid-9fb7aea1-0c01-4527-8def-242f3cde7dc6/conjur"
 
-# both old-style 'conjur' and new-style 'cyberark/conjur'
-INTERNAL_IMAGES=("$CONJUR_REGISTRY/conjur" "$CONJUR_REGISTRY/$IMAGE_NAME")
+for arg in "$@"; do
+  case $arg in
+    --edge )
+      PUBLISH_EDGE=true
+      shift
+      ;;
+    --registry-prefix=* )
+      REGISTRY_PREFIX="${arg#*=}"
+      shift
+      ;;
+    --registry-prefix )
+      echo "This script expects the flag value to be provided as --registry-prefix=STRING"
+      print_help
+      exit 1
+      ;;
+    * )
+      echo "Unknown option: ${arg}"
+      print_help
+      exit 1
+      ;;
+    esac
+done
 
-function main() {
-  # Push the VERSION-SHA tag to our internal registry
-  tag_and_push "$TAG" "$SOURCE_IMAGE" "${INTERNAL_IMAGES[@]}"
+if [[ "${PUBLISH_EDGE}" = true ]]; then
 
-  # Push the latest tag to our internal registry
-  tag_and_push latest "$SOURCE_IMAGE" "${INTERNAL_IMAGES[@]}"
+  if [[ ! -z "${REGISTRY_PREFIX}" ]]; then
+    REGISTRY_PREFIX="${REGISTRY_PREFIX}/"
+  fi
 
-  # Push 1-stable and 1.x-stable for 1.x.y tag to our internal registry
-  # Strip the "v" prefix from the tag name before computing major / minor
-  # tag versions.
-  local prefix_versions
-  readarray -t prefix_versions < <(gen_versions "${TAG_NAME//"v"}")
-  for v in "${prefix_versions[@]}"; do
-    tag_and_push "$v-stable" "$SOURCE_IMAGE" "${INTERNAL_IMAGES[@]}"
-  done
+  # This script is running to publish the edge tagged image to DockerHub
+  tag_and_push "edge" "${LOCAL_IMAGE}" "${REGISTRY_PREFIX}${IMAGE_NAME}"
 
-  # Strip the "v" prefix from the tag name and push
-  # latest, 1.x.y, 1.x, and 1 to DockerHub
-  for v in latest "${TAG_NAME//"v"}" "${prefix_versions[@]}"; do
-    tag_and_push "$v" "$SOURCE_IMAGE" "$IMAGE_NAME"
+elif [[ ! -z "${TAG_NAME:-}" ]]; then
+
+  # This is running on a tag-triggered build, and public images should be published
+  TAG="${TAG_NAME//"v"}"
+
+  # Push latest, 1.x.y, 1.x, and 1 to DockerHub
+  readarray -t prefix_versions < <(gen_versions "${TAG}")
+  for v in latest "${TAG}" "${prefix_versions[@]}"; do
+    tag_and_push "${v}" "${LOCAL_IMAGE}" "${IMAGE_NAME}"
   done
 
   # Publish only the tag version to the Redhat container registry
-  if docker login scan.connect.redhat.com -u unused -p "$REDHAT_API_KEY"; then
-    tag_and_push "${TAG_NAME//"v"}" "$RH_SOURCE_IMAGE" "$REDHAT_IMAGE"
+  if docker login scan.connect.redhat.com -u unused -p "${REDHAT_API_KEY}"; then
+    tag_and_push "${TAG}" "${RH_LOCAL_IMAGE}" "${REDHAT_IMAGE}"
   else
     echo 'Failed to log in to scan.connect.redhat.com'
     exit 1
   fi
-}
 
-# tag_and_publish tag source image1 image2 ...
-function tag_and_push() {
-  local tag="$1"; shift
-  local source="$1"; shift
+elif [[ ! -z "${REGISTRY_PREFIX}" ]]; then
 
-  for image in "$@"; do
-    local target=$image:$tag
-    echo "Tagging and pushing $target..."
-    docker tag "$source" "$target"
-    docker push "$target"
-  done
-}
+  # This is not running on a tag-triggered build, and a registry prefix has
+  # been supplied. Publish to the specified registry.
 
-main
+  # Push the VERSION-SHA tagged images to our internal registry
+  tag_and_push "${TAG}" "${LOCAL_IMAGE}" "${REGISTRY_PREFIX}/conjur"
+  tag_and_push "${TAG}" "conjur-test:${TAG}" "${REGISTRY_PREFIX}/conjur-test"
+  tag_and_push "${TAG}" "conjur-ubi:${TAG}" "${REGISTRY_PREFIX}/conjur-ubi"
+
+else
+
+  echo "This script is not running in a tag-triggered build."
+  echo "Please supply either:"
+  echo " - The registry prefix with the --registry-prefix flag, or"
+  echo " - The --edge flag to publish an edge release to DockerHub."
+  exit 1
+fi